[REFACTOR] #1 : RE-packaging
