+++ app/src/main/AndroidManifest.xml
@@ -10,11 +10,11 @@
         android:supportsRtl="true"
         android:theme="@style/Theme.SOPT_Seminar_30th">
         <activity
-            android:name=".ui.auth.SignUpActivity"
+            android:name=".presentation.ui.auth.SignUpActivity"
             android:exported="false"
             android:windowSoftInputMode="adjustResize" />
         <activity
-            android:name=".ui.auth.SignInActivity"
+            android:name=".presentation.ui.auth.SignInActivity"
             android:exported="true"
             android:windowSoftInputMode="adjustResize">
             <intent-filter>
