+++ app/src/main/java/co/kr/sopt_seminar_30th/presentation/ui/auth/SignUpActivity.kt
@@ -1,10 +1,10 @@
-package co.kr.sopt_seminar_30th.ui.auth
+package co.kr.sopt_seminar_30th.presentation.ui.auth
 
 import android.content.Intent
 import android.os.Bundle
 import co.kr.sopt_seminar_30th.R
 import co.kr.sopt_seminar_30th.databinding.ActivitySignUpBinding
-import co.kr.sopt_seminar_30th.ui.base.BaseActivity
+import co.kr.sopt_seminar_30th.presentation.ui.base.BaseActivity
 
 class SignUpActivity : BaseActivity<ActivitySignUpBinding>() {
     override val TAG: String
